requirements.txt
certifi==2017.4.17
future==0.16.0
peewee==2.10.1
python-telegram-bot==6.1.0
